fountain cold review round 2: correct the usize::MAX doc rationale + 32-bit overflow
R2: 6 fresh blind reviewers, 0 CRITICAL, 0 MODERATE across both leaves (2.1M+
more honest fuzz trials, 0 wrong-bytes-on-success; seals held vs rustc). Two
fountain LOWs fixed (hypertree's lone LOW is a defensible non-defect, left):

- Two reviewers independently flagged that my R1 doc clause 'a real caller cannot
  reach it' is false for decode(), whose k is a free usize parameter (not slice-
  derived like symbol's). Reworded: an absurd k is a caller error caught by a
  clean guarded panic (not UB); a well-formed caller derives k from real data.
- robust_soliton_cdf computed d*(d-1) in usize, overflowing on 32-bit targets at
  k >= 65537. Compute the product in f64 (harmless on 64-bit, correct everywhere).

Gates: fountain 10+2, workspace clippy/fmt/rustdoc -D warnings clean.

+++ fountain-types/src/lt.rs
@@ -54,7 +54,9 @@ pub fn robust_soliton_cdf(k: usize) -> Vec<f64> {
                                                                                   // ideal soliton
     w[1] += 1.0 / kf;
     for (d, wd) in w.iter_mut().enumerate().skip(2) {
-        *wd += 1.0 / ((d * (d - 1)) as f64);
+        // Compute the product in f64 to avoid a usize overflow of `d*(d-1)` on
+        // 32-bit targets when k is large (harmless on 64-bit, but correct everywhere).
+        *wd += 1.0 / ((d as f64) * ((d - 1) as f64));
     }
     // robust correction tau
     for (d, wd) in w.iter_mut().enumerate().take(spike).skip(1) {
